deep_research_agent/requirements.txt
langgraph
langchain-core # Often needed by langgraph components
tavily-python
python-dotenv # Useful for managing environment variables like API keys
streamlit
